requirements.txt
fastapi[all]
sqlalchemy
psycopg
psycopg[binary]
alembic
python-dotenv
pytest
pytest-asyncio
pytest-cov
pytest-mock
pydantic
black
mypy
httpx
isort
sqlmodel
